+++ example/lib/shared/interface/bar_life_controller.dart
@@ -41,5 +41,5 @@ class BarLifeController extends StateController<BarLifeComponent> {
   }
 
   @override
-  void update(double dt) {}
+  void update(double dt, BarLifeComponent component) {}
 }